Reorganize util locations
