+++ plugins/src/client/plugin.rs
@@ -34,7 +34,7 @@ where
         let s = CString::from_raw(c);
         let CreateArgs::<Self> {
             arguments,
-            context: ContextualArguments { network, amount },
+            context: ContextualArguments { network, amount, effects},
         } = serde_json::from_slice(s.to_bytes())?;
         // TODO: Get The wasm ID here?
         // TODO: In theory, these trampoline bounds are robust/serialization safe...
@@ -46,7 +46,7 @@ where
             /// TODO: Carry context's path
             vec![Arc::new("PLUGIN_TRAMPOLINE".into())],
             /// TODO: load database?
-            Arc::new(MapEffectDB::default()),
+            Arc::new(effects)
         );
         let converted = Self::ToType::try_from(arguments)?;
         let compiled = converted.compile(ctx)?;
