+++ sapio/src/contract/abi/continuation.rs
@@ -6,78 +6,38 @@
 
 //! ABI for contract resumption
 
-use crate::util::reverse_path::ReversePath;
+use sapio_base::reverse_path::ReversePath;
 use schemars::schema::RootSchema;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
+use sapio_base::serialization_helpers::SArc;
 /// Instructions for how to resume a contract compilation at a given point
 #[derive(Serialize, Deserialize, JsonSchema, Clone, Debug, PartialEq)]
 pub struct ContinuationPoint {
     /// The arguments required at this point
     /// TODO: De-Duplicate repeated types?
-    pub schema: Option<rs::SArc<RootSchema>>,
+    pub schema: Option<SArc<RootSchema>>,
     /// The path at which this was compiled
-    #[serde(serialize_with = "rs::serializer")]
-    #[serde(deserialize_with = "rs::deserializer")]
+    #[serde(serialize_with = "sapio_base::serialization_helpers::serializer")]
+    #[serde(deserialize_with = "sapio_base::serialization_helpers::deserializer")]
     pub path: Arc<ReversePath<String>>,
 }
 impl ContinuationPoint {
     /// Creates a new continuation
     pub fn at(schema: Option<Arc<RootSchema>>, path: Arc<ReversePath<String>>) -> Self {
         ContinuationPoint {
-            schema: schema.map(rs::SArc),
+            schema: schema.map(SArc),
             path,
         }
     }
 }
 
-/// Helpers for serializing Arcs
-pub mod rs {
-    use schemars::JsonSchema;
-    use serde::{Deserialize, Deserializer, Serialize, Serializer};
-    use std::borrow::Borrow;
-    use std::sync::Arc;
 
-    /// Serializable Arc Type
-    #[derive(Serialize, Deserialize, JsonSchema, Clone, Debug, PartialEq, PartialOrd, Eq, Hash)]
-    #[serde(
-        bound = "T: Serialize + for<'d> Deserialize<'d> + JsonSchema + std::fmt::Debug + Clone "
-    )]
-    #[serde(transparent)]
-    pub struct SArc<T>(
-        #[serde(serialize_with = "serializer")]
-        #[serde(deserialize_with = "deserializer")]
-        pub Arc<T>,
-    );
-    /// arc serializer
-    pub fn serializer<T, S>(v: &Arc<T>, s: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-        T: Serialize,
-    {
-        let b: &T = v.borrow();
-        b.serialize(s)
-    }
-    /// arc deserializer
-    pub fn deserializer<'de, T, D>(d: D) -> Result<Arc<T>, D::Error>
-    where
-        D: Deserializer<'de>,
-        T: Deserialize<'de>,
-    {
-        Ok(Arc::new(T::deserialize(d)?))
-    }
-}
 
 #[cfg(test)]
 mod test {
     use super::*;
-    #[test]
-    fn test_sarc_ser() -> Result<(), Box<dyn std::error::Error>> {
-        assert_eq!(serde_json::to_string(&rs::SArc(Arc::new(1)))?, "1");
-        Ok(())
-    }
-
     #[test]
     fn test_continuation_point_ser() -> Result<(), Box<dyn std::error::Error>> {
         let a: ContinuationPoint = ContinuationPoint::at(
